Version 59.0.0
Version: 59.0.0
Timestamp: 2026-02-07T16:46:35.470287

+++ Resources/Texture2D Blit.shader
@@ -26,7 +26,7 @@ Shader "Oculus/Texture2D Blit" {
                 struct v2f
                 {
                     float4 vertex : SV_POSITION;
-                    half2 texcoord : TEXCOORD0;
+                    float2 texcoord : TEXCOORD0;
                 };
 
                 sampler2D _MainTex;
